requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.0.3
openpyxl==3.1.2
qrcode==7.4.2
Pillow==10.0.0
werkzeug==2.3.7
python-dateutil==2.8.2